requirements.txt
albumentations>=1.4.0
asyncpg>=0.29.0
fastapi>=0.110.0
joblib>=1.3.2

numpy>=1.26.0,<2.0
pillow>=10.2.0
psycopg2-binary>=2.9.9
pyjwt>=2.8.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

pytorch-metric-learning>=2.4.0
scikit-learn>=1.4.0
seaborn>=0.13.0
sqlalchemy>=2.0.25

torch
torchvision

ultralytics>=8.2.0
uvicorn>=0.27.0
